backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "micro-saas-backend"
version = "0.1.0"
description = "FastAPI Lambda API (Serverless)"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0,<2",
  "bcrypt>=5.0.0",
  "fastapi>=0.115.0,<1",
  "mangum>=0.19.0,<1",
  "passlib[bcrypt]>=1.7.4",
  "psycopg2-binary>=2.9.9,<3",
  "pyjwt>=2.8.0",
  "sqlalchemy>=2.0.0,<3",
  "sqlmodel>=0.0.22,<1",
]

[dependency-groups]
dev = [
  "httpx>=0.27.0,<1",
  "pytest>=8.0.0,<9",
  "ruff>=0.8.0,<1",
  "uvicorn>=0.32.0,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
